refactor(mocks): extract judge module from mockApi (#991 M3)
Move reviews/pendingJudges state and the judge-backlog reimplementation (bucketOf…computeBacklog) plus the 8 getRunReview…rerunJudge methods into judge.ts as judgeApi; index.ts spreads it, imports reviews for the findings methods, and re-exports the fixture-facing internals.
